requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dataclasses-json==0.6.1
exceptiongroup==1.1.3
ez-setup==0.9
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langchain-experimental==0.0.36
langsmith==0.0.53
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.0
openai==0.28.1
packaging==23.2
psycopg2==2.9.6
psycopg2-binary==2.9.9
pydantic==2.4.2
pyscopg2==66.0.2
python-dotenv==1.0.0
python-environ==0.4.54
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
yarl==1.9.2